Implemented auto-escaping of variable output in templates
Per default all variables are now escaped. This behaviour can be turned off
in the template. Autoescaping puts the burden of thinking about escaping on the
filter writes. Filters and templates can and should be written so that they
work both in auto-escaped and non-autoescaped environments.

I added a lot of new filters with unit tests in this commit to test the autoescape
feature.

Guideline for writing filters
-------------------------------

  * decorate the filter with `isSafe` if the filter does not change the safeness
  of the string (e.g. a string passed to `lower` will be as safe as before).
  In practice this means that your filter is safe if it does not introduce
  HTML-unsafe characters (<, >, ', " or &) that were not already present and
  if you are not removing characters, which might unbalance HTML tags.

It is NOT enough to just mark the filter as `isSafe` you must make sure
it actually IS safe. As an introductory exercise think about why - in contrast
to `lower` - the `upper` filter is NOT safe.

Also note that marking a filter as `isSafe` will coerce the filter's return value
to a string!

  * `markSafe(output)` mark the output of your filter as safe if the output is
   ensured to be safe independant of the safeness of the input (if you manually
   deal with the safeness in the filter with markSafe, then you don't have to
   `isSafe=true` the filter as that would be redundant but not dangerous).

Also see: <https://docs.djangoproject.com/en/dev/howto/custom-template-tags/#filters-and-auto-escaping>

(before jumping into the django docs you'll have to read the 'Notable differences'
section below)

Guidelines to writing tags
-------------------------------

  * the output of a tag (if it has any) is NOT automatically run through the auto-
  escaping filters
  * if you store results in a context variable, make sure you `markSafe()`
  where appropriate

Also see: <https://docs.djangoproject.com/en/dev/howto/custom-template-tags/#auto-escaping-considerations>

"needsAutoescape" filter decoration
-------------------------------------

This is relevant for filters which add html to their input. Such filters
must figure out whether they should escape their input or not before adding
the html. Whether to escape the input or not depends on 1) the current autoescape
setting and 2) the safeness of the input.

If a filter needs to know about the current autoescape setting within which
it is invoked, it needs the decoration `needsAutoescape=true`. The filter will
then be invoked with an additional argument which designates the current auto-
escape setting.

For example, look at the `filter/htmlstrings/linebreaks` filter: if this filter
is invoked in an autoescape=true environment but its input is not yet safe,
it has to escape its input before wrapping it in additional html. But if
the filter is invoked in autoescape=false or if the input is already safe
it should not do any escaping.

Noteable differences to django
----------------------------------

We don't have a special `String` class which is safe. Instead I coerce a string
primitve into a String instance to be able to attach the property `isSafe=true`
to it.

 * to test for safeness of a string use `utils.isSafe(string)`
 * to test for escape need of a string use `utils.isMarkedForEscaping(string)`
  (markedForEscaping is necessary if you want to force escaping for output
  even if in a non-autoescaped environment)

 * What should happen if you escape a non-string? Django calls __repr__ and
 escapes that string. I try to call `toString()` and if that doesn't work
 I stringify the value with `JSON.stringify(value)` before escaping it.

 * (only relevant if you want to understand the implementation) Look at
 IfEqual.prototype.render: I am comparing two variables. Before comparision I
 have to convert the variables `toString()` if they happen to be instanceof String
 because the string primitive is not strict equal to the String instance.
 `"test" !== new String('test')`. This is likely to bite us in other places as well.

== related django discussion ==

in order of relevance

<https://code.djangoproject.com/wiki/AutoEscaping>

after autoescape commit
<https://groups.google.com/d/topic/django-developers/luUW1ZFe4zQ/discussion>

autoescape patch discussion
<https://groups.google.com/d/topic/django-developers/fK64bAS4HxA/discussion>
<https://code.djangoproject.com/ticket/2359>

escaping in templates, 2007
<https://groups.google.com/d/topic/django-developers/IGsLpBwiKbc/discussion>

2006 propsal autoescape
<https://groups.google.com/d/topic/django-developers/F9Hf7NZ4ZKs/discussion>

after oscon
<https://groups.google.com/d/topic/django-developers/hnwSztQe5Sk/discussion>

is_safe problem
<https://groups.google.com/d/topic/django-developers/1wlzLyP_MM4/discussion>

global escape
<https://groups.google.com/d/topic/django-users/IdqInsucY90/discussion>

storing news articles - retaining some HTML
<https://groups.google.com/d/topic/django-users/E8-CGNOhiq0/discussion>

templates and html escaping
<https://groups.google.com/d/topic/django-developers/5Ei73UBCaRU/discussion>