add {% loadfilter %} and {% loadtag %}
two different functions because I don't want to always decorate the functions
just to designate them as filter or tag; so the developer has to split them into
seperate modules and load them with the right tag.

the tests are very artifical since usually you'd load the tag/filter module
form a package (and not with a relative path).