add filters `key` and `byKey`
if you want to iterate over the properties of an object in a {% for %} loop
you'r gonna need those - see new unit tests.

this is all a bit akward, imo, because in JS `for..in` and `for each ... in` already
has semantics and those are different from python's. so those two filters are
more a clutch until we figure out if and how we want to deviate from how django
templates do a `for` loop.